urngd: Add support for read()ing entropy
This allows us to attach a hwrng, for example.

At most one sample will be taken every time we also add entropy via the
jitter mechanism.  We won't try reading if the stream hasn't indicated
its readiness, and we'll go back to waiting if ever the stream produces
0 bytes or an error on read.  If the read was successful, we will
optimistically assume it will be next time, too, avoiding the need to
asynchronously wait for sample availability most of the time.

Signed-off-by: Nathaniel Wesley Filardo <nwfilardo@gmail.com>